requirements.txt
requests>=2.31.0
selenium>=4.15.0
tqdm>=4.66.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
m3u8>=4.0.0
pycryptodome>=3.19.0
mangum>=0.17.0
